fix: Create new thread_rng rather than using cache (#131)
This code is likely to be used in asynchronous environments. The usage
of `rand::thread_rng()` means that the consumer can never be sync and
send, because it uses a RC to cache the ThreadRng. It's only a very
small overhead to create a new context every time, and this will enable
calling code to also be asynchronous.